+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added an option to specify the custom colour for non-player/system messages' title (part before the comma) in chat.
+- Injected a silly game difficulty indicator to the lobby browser, likely phased out by the DAHM 1.16's native lobby ordering.
 
 ## 1.0.2
 - Added a fix for making `menu_waiting_is_not_ready` and `menu_waiting_is_ready` strings take custom colours (see updates to zneix's user.lua).